backend/model/requirements.txt
fastapi
uvicorn
torch
torchvision
timm
pillow
numpy
python-multipart
